requirements.txt
streamlit>=1.28.0
Pillow>=9.0.0
pandas>=1.5.0
openpyxl>=3.9.0
streamlit-cropper>=0.2.0
beautifulsoup4>=4.11.0
requests>=2.28.0
selenium>=4.0.0
# UTF-8 encoding support
chardet>=4.0.0

